+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v0.36.9 (WIP)
+## v0.36.9
 
 - Updated the Discord `AuthUser.Name` field to use `global_name` ([#7603](https://github.com/pocketbase/pocketbase/pull/7603); thanks @HansHans135).
 
@@ -6,7 +6,7 @@
 
 - Added extra OAuth2 checks when downloading the avatar URL to prevent internal network probing requests in case of a malicious/vulnerable vendor.
 
-- Updated `modernc.org/sqlite` to v1.48.2 _(vfs and other minor fixes)_.
+- Updated `modernc.org/sqlite` to v1.48.2 _(vfs and other error path related fixes)_.
 
 - Updated min Go GitHub action version to 1.26.2 because it comes with some [minor security fixes](https://github.com/golang/go/issues?q=milestone%3AGo1.26.2).